backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary==2.9.10
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
Pillow==11.1.0
drf-yasg==1.21.8
python-decouple==3.8
anthropic==0.43.0
openai==1.58.1
